Refactored code to be common between all Auth0 api key calls
