+++ apps/raptor/test/integration/raptor_web/controllers/api_key_controller_test.exs
@@ -83,8 +83,8 @@ defmodule Raptor.ApiKeyControllerTest do
       Raptor.Services.Auth0UserDataStore.persist(auth0_user_data)
 
       # Do not allow Auth0 calls
-      allow(Auth0Management.get_users_by_api_key(any()),
-        return: {:error, "error"}
+      allow(Tesla.get(any(), any()),
+        return: {:error, nil}
       )
 
       {:ok, response} =
